backend/requirements.txt
# PolicyPulse 后端依赖包
# 路径: /mnt/okcomputer/output/backend/requirements.txt

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
motor==3.3.2
pymongo==4.6.0
redis==5.0.1

# 异步支持
aiohttp==3.9.1
aiofiles==23.2.1

# AI服务
openai==1.3.7
tiktoken==0.5.1

# 数据处理
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2

# 安全认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 邮件服务
aiosmtplib==3.0.1
email-validator==2.1.0

# 短信服务
alibabacloud-dysmsapi20170525==2.0.24

# 任务队列
celery==5.3.4

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 日志和监控
structlog==23.2.0
prometheus-client==0.19.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# 其他工具
click==8.1.7
jinja2==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0

# 生产环境
gevent==23.9.1
gunicorn==21.2.0